+++ app/models/phone.rb
@@ -1,3 +1,7 @@
 class Phone < ApplicationRecord
-    has_many :claims
+    has_many :claims, dependent: :destroy
+
+    validates :brand, presence: true
+    validates :model, presence: true
+    validates :price, presence: true
 end